center component created
