+++ .eslintcache
@@ -1,3 +1,1 @@
-
-[{"C:\\Users\\Kingsley\\VoomswayCompany\\TwitterClone\\twiiter-clone\\src\\component\\LeftComponent.jsx":"1"},{"size":478,"mtime":1611312181565,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"xj5af6","C:\\Users\\Kingsley\\VoomswayCompany\\TwitterClone\\twiiter-clone\\src\\component\\LeftComponent.jsx",[]]
-
+[{"C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\index.js":"1","C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\App.js":"2","C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\reportWebVitals.js":"3","C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\component\\LandingPage.jsx":"4","C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\component\\LeftComponent.jsx":"5","C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\component\\centerComponent\\CenterPage.jsx":"6","C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\component\\centerComponent\\Head.jsx":"7","C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\component\\centerComponent\\Tweet.jsx":"8"},{"size":517,"mtime":1611238194963,"results":"9","hashOfConfig":"10"},{"size":179,"mtime":1611310941456,"results":"11","hashOfConfig":"12"},{"size":375,"mtime":1611238194963,"results":"13","hashOfConfig":"12"},{"size":463,"mtime":1611310941450,"results":"14","hashOfConfig":"10"},{"size":478,"mtime":1611312422960,"results":"15","hashOfConfig":"10"},{"size":376,"mtime":1611323243977,"results":"16","hashOfConfig":"12"},{"size":1907,"mtime":1611323023842,"results":"17","hashOfConfig":"12"},{"size":874,"mtime":1611324671898,"results":"18","hashOfConfig":"12"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"21"},"1458yci",{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1eilnnp",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"21"},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\index.js",[],["36","37"],"C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\App.js",[],"C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\reportWebVitals.js",[],"C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\component\\LandingPage.jsx",[],"C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\component\\LeftComponent.jsx",[],"C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\component\\centerComponent\\CenterPage.jsx",[],"C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\component\\centerComponent\\Head.jsx",[],"C:\\Users\\voomsway\\Documents\\workspace\\web\\twitter-clone\\src\\component\\centerComponent\\Tweet.jsx",[],{"ruleId":"38","replacedBy":"39"},{"ruleId":"40","replacedBy":"41"},"no-native-reassign",["42"],"no-negated-in-lhs",["43"],"no-global-assign","no-unsafe-negation"]
