+++ src/component/centerComponent/CenterPage.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import '../style.css';
+import Head from './Head';
+import Tweet from './Tweet';
+
+
+const CenterComponent = () => {
+    return(
+        <>
+        <div className="centerComponent">
+         <Head />
+         <br/>
+         <br />
+         <Tweet />
+           
+        </div>
+
+        </>
+    )
+}
+
+
+export default CenterComponent;
