+++ src/component/centerComponent/Tweet.jsx
@@ -1,0 +1,35 @@
+
+
+const Tweet = () => {
+    return(
+        <div className="jumbotron">
+       <div className="row">
+            <div className="col-2">
+                {/* <img style={{borderRadius: '50%'}} src="https://assets.telegraphindia.com/telegraph/2020/Oct/1603223670_google.jpg" alt="" /> */}
+            </div>
+
+            <div className="col-9">
+                {/* Top inner row */}
+
+               <div className="row">
+                <div className="col-4">
+                <small>Google International</small>
+                </div>
+                <div className="col-2">
+                <small>Google</small>
+                </div>
+                <div className="col-2">
+                    <small>14m</small>
+                </div>
+
+               </div>
+
+            </div>
+
+       </div>
+
+      </div>
+    );
+}
+
+export default Tweet;
